pyproject.toml
[project]
name = "acf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.2",
    "channels>=4.3.2",
    "daphne>=4.2.1",
    "django>=6.0.2",
    "djlint>=1.36.4",
    "google-genai>=1.65.0",
    "httpx>=0.28.1",
    "pypdf>=6.7.4",
    "python-docx>=1.2.0",
    "redis>=7.2.1",
]
